Refactor but-db
That way the overall structure becomes clearer
