+++ crates/but-db/src/table/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) mod butler_actions;
+pub(crate) mod ci_checks;
+pub(crate) mod claude;
+pub(crate) mod file_write_locks;
+pub(crate) mod forge_reviews;
+pub(crate) mod gerrit_metadata;
+pub(crate) mod hunk_assignments;
+pub(crate) mod workflows;
+pub(crate) mod workspace_rules;
